added authorization validation on api

+++ backend/src/main/java/com/ku/quizzical/app/controller/question/QuestionController.java
@@ -10,25 +10,41 @@ import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RestController;
+import com.ku.quizzical.app.controller.quiz.Quiz;
+import com.ku.quizzical.app.controller.user.User;
+import com.ku.quizzical.app.controller.quiz.QuizRepository;
+import com.ku.quizzical.app.helper.AuthorizationValidationHelper;
+import com.ku.quizzical.app.helper.BackendValidationHelper;
 
 @CrossOrigin
 @RestController
 public final class QuestionController {
     // Instance Fields
     private QuestionDatabaseService service;
+    private QuizRepository quizRepository;
+    private QuestionRepository repository;
 
     // Constructor Method
-    public QuestionController(QuestionDatabaseService service) {
+    public QuestionController(QuestionDatabaseService service, QuizRepository quizRepository,
+            QuestionRepository repository) {
         super();
         this.service = service;
+        this.quizRepository = quizRepository;
+        this.repository = repository;
     }
 
     // CREATE Method
     // Saves a Question
     @PostMapping("/quizzes/{quizId}/questions")
     public ResponseEntity<QuestionDto> saveQuestion(@PathVariable String quizId,
+            @RequestHeader("Authorization") String authorizationHeader,
             @RequestBody QuestionDto question) {
+        Quiz matchingQuiz = BackendValidationHelper.validateExistingResourceWithFallthrough(
+                "Question Quiz", question.quizId(), this.quizRepository::findById);
+        AuthorizationValidationHelper.validateAuthorization(authorizationHeader,
+                matchingQuiz.getUserId());
         return new ResponseEntity<QuestionDto>(this.service.saveQuestion(quizId, question),
                 HttpStatus.OK);
     }
@@ -52,7 +68,13 @@ public final class QuestionController {
     // Updates a Question
     @PatchMapping("/quizzes/{quizId}/questions/{id}")
     public ResponseEntity<QuestionDto> updateQuestion(@PathVariable("quizId") String quizId,
-            @PathVariable("id") String id, @RequestBody QuestionUpdateRequest question) {
+            @PathVariable("id") String id,
+            @RequestHeader("Authorization") String authorizationHeader,
+            @RequestBody QuestionUpdateRequest question) {
+        Question matchingQuestion = BackendValidationHelper
+                .validateExistingResourceWithFallthrough("Question", id, this.repository::findById);
+        AuthorizationValidationHelper.validateAuthorization(authorizationHeader,
+                matchingQuestion.getUserId());
         return new ResponseEntity<QuestionDto>(this.service.updateQuestion(quizId, id, question),
                 HttpStatus.OK);
     }
@@ -61,12 +83,13 @@ public final class QuestionController {
     // Deletes a Question
     @DeleteMapping("/quizzes/{quizId}/questions/{id}")
     public String deleteQuestion(@PathVariable("quizId") String quizId,
-            @PathVariable("id") String id) {
-        try {
-            this.service.deleteQuestion(quizId, id);
-        } catch (Exception e) {
-            // Do Nothing
-        }
+            @PathVariable("id") String id,
+            @RequestHeader("Authorization") String authorizationHeader) {
+        Question matchingQuestion = BackendValidationHelper
+                .validateExistingResourceWithFallthrough("Question", id, this.repository::findById);
+        AuthorizationValidationHelper.validateAuthorization(authorizationHeader,
+                matchingQuestion.getUserId());
+        this.service.deleteQuestion(quizId, id);
         return "\"The Question with id \\\"" + id + "\\\" has been deleted.\"";
     }
 }
